+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Notable changes of tdmclient. Release versions refer to [https://pypi.org/projec
 
 - Tool `info` to display the variables, events and native functions of a robot.
 - Existence of a variable in a node can be tested with `in` or `not in`.
+- In tool `list`, the product id (content of the node variable `_productId`) is displayed if it's available, with its meaning if it's known. 
 
 ### Fixed